вынесена загрузка и добавлен алерт
