--- homework-7/homework-7/ImageLoaderDelegate.swift
@@ -1,16 +1,0 @@
-//
-//  ImageLoaderDelegate.swift
-//  homework-7
-//
-//  Created by Денис Гынгазов on 02.12.2020.
-//
-
-import UIKit
-
-final class TableViewDelegate: NSObject, UITableViewDelegate {
-	
-	func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-
-		return 50
-	}
-}
